fixed destroy link

+++ db/migrate/20141027221053_create_products.rb
@@ -3,9 +3,9 @@ class CreateProducts < ActiveRecord::Migration
     create_table :products do |t|
       t.string :name
       t.string :asin
-      t.integer :price
-      t.integer :regPrice
-      t.integer :rate_of_change
+      t.decimal :price
+      t.decimal :regPrice
+      t.decimal :rate_of_change
       t.string :description
       t.string :url
       t.string :img_url
